refactor: move common selectors to avoid duplication and circular imports
